+++ Plugin/src/Content/Enemies/ReadyJP.cs
@@ -1,0 +1,8 @@
+namespace CodeRebirth.src.Content.Enemies;
+public class ReadyJP : CodeRebirthEnemyAI // made for future implementation
+{
+    public override void Start() 
+    {
+        base.Start();
+    }
+}