+++ Plugin/src/Content/Items/Money.cs
@@ -1,4 +1,3 @@
-using System;
 using CodeRebirth.src.Util.Extensions;
 using Unity.Netcode;